Only show lights that are 'visible' in the entity selection
When lights are grouped, the individual bulbs are hidden but they still show up in the selection list.  For grouped lights with the individuals hidden, we dont want those to be selected (as they're hidden) in favor of using the groups.    This also allows for easier replacement of failed bulbs as we can just add them back to the group and not have to go edit every scene, automation, and adaptive lighting configuration.